Streamlit/requirements.txt
streamlit
requests
opencensus-ext-azure